made appliedFIlters
