+++ src/pages/Explore/components/Body/components/Filters/components/PriceFilter/styles.module.scss
