+++ src/pages/Explore/components/Body/components/Filters/styles.module.scss
@@ -1,5 +1,17 @@
 @import 'src/styles/helpers.scss';
 
+.filtersLabels {
+  padding: 16px 0 24px;
+}
+
+.sorting {
+  padding: 26px 0;
+  @include flexbox(center, space-between);
+
+  &Dropdown {
+    min-width: 200px;
+  }
+}
 .filters {
   padding: 24px;
   background-color: rgba($color-light-4, 0.4);
@@ -15,9 +27,13 @@
 }
 
 .dropdown {
-  width: 200px;
+  min-width: 150px;
 
   margin-right: 16px;
+
+  &Small{
+    min-width: 150px;
+  }
 }
 
 .prices {
