+++ src/pages/Explore/components/Body/components/index.ts
@@ -1,0 +1,4 @@
+
+export { default as Filters } from './Filters';
+export { default as Labels } from './Labels';
+export { default as FilterLabel } from './FilterLabel';
