+++ src/pages/Explore/components/Title/styles.module.scss
@@ -1,0 +1,6 @@
+@import '/src/styles/helpers.scss';
+
+.title {
+  margin-top: 92px;
+  margin-bottom: 88px;
+}
