+++ src/pages/Explore/components/index.ts
@@ -1,4 +1,3 @@
 
-export { default as Filters } from './Filters';
-export { default as Labels } from './Labels';
-export { default as FilterLabel } from './FilterLabel';
+export { default as Title } from './Title';
+export { default as Body } from './Body';
