+++ src/pages/Explore/styles.module.scss
@@ -6,46 +6,3 @@
     padding: 0;
   }
 }
-
-.container {
-  @include container;
-}
-
-.title {
-  margin-top: 92px;
-  margin-bottom: 88px;
-}
-
-.categories {
-  @include flexbox(flex-start, center);
-  .category + .category {
-    margin-left: 16px;
-  }
-}
-
-.filters {
-  padding: 16px 0 24px;
-}
-
-.sorting {
-  padding: 26px 0;
-  @include flexbox(center, space-between);
-
-  &Dropdown {
-    min-width: 200px;
-  }
-}
-
-.cards {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(344px, 1fr));
-
-  gap: 24px;
-
-  margin-bottom: 48px;
-}
-
-.load{
-  @include flexbox(center, center);
-  margin-bottom: 80px;
-}
